requirements .txt
streamlit
pillow
python-dotenv
ipython
google-generativeai 